requirements.txt
# Core
numpy>=1.24
soundfile>=0.12
scipy>=1.11
torch>=2.6
requests>=2.32
matplotlib>=3.8
python-docx>=1.1

# Model (Diffusers text-to-audio)
diffusers>=0.32
transformers>=4.45,<5
accelerate>=0.34
safetensors>=0.4
huggingface_hub>=0.24

# Optional UI
gradio>=4.0
pywebview>=5.0

# Windows desktop backend (pywebview WinForms) depends on pythonnet.
# Pin these together to avoid mismatched Python.Runtime.dll vs loader code in packaged builds.
pythonnet==3.0.3
clr-loader==0.2.7
